Move added packages to util/math
